REFACTOR: move playback playlist to playlist.rs
